+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { UserService } from './services/user.service';
-import { SEOService } from './services/seo.service';
 import { BotsService } from './services/bots.service';
+import { AuthService } from './services/auth.service';
 
 @Component({
   selector: 'app-root',
@@ -10,11 +10,14 @@ import { BotsService } from './services/bots.service';
 })
 export class AppComponent implements OnInit {
   constructor(
+    private auth: AuthService,
     private botService: BotsService,
     private userService: UserService) {}
 
   async ngOnInit() {
     await this.botService.init();
+
+    this.auth.validateKey();
 
     await this.userService.updateUser();
     await this.userService.updateSavedUser();