+++ src/app/bot-preview/bot-preview.component.spec.ts
@@ -1,6 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { BotPreviewComponent } from './bot-preview.component';
+import { By } from '@angular/platform-browser';
 
 describe('BotPreviewComponent', () => {
   let component: BotPreviewComponent;
@@ -21,5 +22,14 @@ describe('BotPreviewComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('click delete button, calls delete', () => {
+    // const spy = spyOn(component, 'delete');
+    const el = fixture.debugElement.query(By.css('#delete')).nativeElement;
+
+    el.click();
+
+    // expect(spy).toHaveBeenCalled();
   });
 });