+++ src/app/member-username/member-username.component.ts
@@ -6,6 +6,6 @@ import { Component, Input } from '@angular/core';
   styleUrls: ['./member-username.component.css']
 })
 export class MemberUsernameComponent {
-  @Input() member;
+  @Input() user;
   @Input() withAvatar = true;
 }